requirements.txt
Flask>=2.0
flask-cors>=4.0
PyMuPDF>=1.24
google-generativeai>=0.5
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-google-genai>=1.0
langchain-text-splitters>=0.2
faiss-cpu>=1.8
pydantic>=2.0
requests
beautifulsoup4
python-dotenv
redis
pydantic
pydantic-settings
typing_extensions
numpy
scikit-learn